Perform operate when pressing equal or chaining
Rewrote the code to be a little more simple with retrieveing the 2nd
number because The Odin Project provided a calculator as an example that
was less complex than the one I orginally wrote with regards to how it
displayed numbers and operators. I'm unclear on if I'm supposed to be
using objects or not for this project but have implemented using an
array to track operation chaining.